ml-services/requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.21.0
pydantic==1.10.7
sqlalchemy==2.0.9
psycopg2-binary==2.9.5
alembic==1.10.2
redis==4.5.4
pandas==2.0.0
numpy==1.24.2
xarray==2023.3.0
netCDF4==1.6.3
h5netcdf==1.1.0
scikit-learn==1.2.2
faiss-cpu==1.7.3
chromadb==0.3.21
langchain==0.0.133
openai==0.27.4
anthropic==0.2.8
sentence-transformers==2.2.2
transformers==4.28.1
torch==2.0.0
plotly==5.14.0
kaleido==0.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.24.0
aiofiles==23.1.0
celery==5.2.7
flower==1.2.0
pytest==7.2.2
pytest-asyncio==0.21.0
black==23.3.0
isort==5.12.0
flake8==6.0.0
mypy==1.2.0